# Use Arc for environment variables on the stack